Add Pipeline

+++ makeDataSet.sh
@@ -13,20 +13,75 @@ function processCommonFromDB() {
     mkdir -p ${result_dir}/${target}
    
     logpath=${log_dir}/${target}.log
-    if [ -f ${logpath} ]
-    then
-		rm ${logpath}
-    fi
 
-    python ${src_dir}/processCommonFromDB.py \
-		   --outpath ${result_dir}/${target}/common_procesed.pkl \
-		   --logpath ${logpath} \
-		   --target ${target}
+	outpath=${result_dir}/${target}/common_procesed.pkl
+	
+	if [ ! -e ${outpath} ];then
+		
+		if [ -f ${logpath} ]
+		then
+			rm ${logpath}
+		fi
+		echo "***processCommonFromDB.py***"
+		python ${src_dir}/processCommonFromDB.py \
+			   --outpath ${outpath} \
+			   --logpath ${logpath} \
+			   --target ${target}
+	fi
 }
 
-function pipe() {
-    processCommonFromDB ${target}  
+function getTarget() {
+
+	outdir=${result_dir}/all/getTarget${datalabel}
+	mkdir -p ${outdir}
+	outpath=${outdir}/title_target.pkl
+	
+	if [ ! -e ${outpath} ];then
+
+		echo "***getTarget.py***"
+		
+		if [ -z "${datalabel}" ];then
+			python ${src_dir}/getTarget.py \
+				   --outpath ${outpath} 
+		else
+			python ${src_dir}/getTarget.py \
+				   --outpath ${outpath} \
+				   --small
+		fi
+
+	fi
 }
+
+function makeDictionary() {
+
+	inpath=${result_dir}/all/getTarget${datalabel}/title_target.pkl
+	outdir=${result_dir}/all/makeDictionary${datalabel}
+	mkdir -p ${outdir}
+	if [ ! -e ${outdir}/df_id.pkl ];then
+
+		echo "***makeDictionary.py***"
+		python -W ignore ${src_dir}/makeDictionary.py \
+			   --inpath ${inpath} \
+			   --outdir ${outdir}
+	fi
+}
+
+function getDataset() {
+
+	inpath=${result_dir}/all/makeDictionary${datalabel}/
+	outdir=${result_dir}/all/dataset${datalabel}/${config_name}
+	mkdir -p ${outdir}
+	if [ ! -e ${outdir}/train.pkl -o ! -e ${outdir}/val.pkl -o ! -e ${outdir}/test.pkl ];then
+
+		echo "***dataset.py***"
+		python -W ignore ${src_dir}/dataset.py \
+			   --inpath ${inpath} \
+			   --outdir ${outdir} \
+			   --config "${config}"
+	fi
+}
+
+datalabel=""
 
 for option in "$@"
 do
@@ -34,6 +89,10 @@ do
 	"-t"|"--target" )
 	    target="${2}"
 	    shift 2
+		;;
+	"-s"|"--small" )
+	    datalabel="_small"
+	    shift 1
     esac
 done
 
@@ -41,9 +100,23 @@ if [ -z "${target}" ]
 then
     for target in `cat ${target_list}`
     do
-	pipe ${target}
+		processCommonFromDB ${target}
     done
 else
-    pipe ${target}
+    processCommonFromDB ${target}
 fi
 
+getTarget
+
+makeDictionary
+
+dataset_config_path="dataset/config.json"
+dataset_config=`cat ${dataset_config_path}`
+length=`echo ${dataset_config} | jq length`
+
+for i in $( seq 1 ${length} )
+do
+	config_name=dataset${i}
+	config=`echo ${dataset_config} | jq -r .${config_name}`
+	getDataset
+done
